+++ my_iot/context.py
@@ -1,17 +1,19 @@
 from __future__ import annotations
 
-from asyncio import create_task
+import asyncio
+from asyncio import CancelledError, create_task, gather, sleep
+from contextlib import suppress
 from dataclasses import InitVar, dataclass, field
 from datetime import datetime, timedelta
 from types import ModuleType
 from typing import Iterable, List, Mapping
 
-from aiohttp import ClientSession
+from aiohttp import ClientConnectorError, ClientSession
 from loguru import logger
 from sqlitemap import Connection
 
 from my_iot.constants import ACTUAL_KEY, HTTP_TIMEOUT
-from my_iot.helpers import setattr_async, timestamp_key
+from my_iot.helpers import run_in_executor, timestamp_key
 from my_iot.routing import EventRouter
 from my_iot.services.base import Service
 from my_iot.types_ import Event
@@ -20,6 +22,7 @@ EVENT_INCLUDE = {'timestamp', 'value'}  # logged value uses optimised representa
 EVENT_EXCLUDE = {'is_logged'}  # is only needed before writing
 
 
+# FIXME: should be named `Runner`.
 @dataclass
 class Context:
     # User-defined automation.
@@ -41,33 +44,55 @@ class Context:
         self.router = getattr(automation, 'router', None) or self.router
         self.services = getattr(automation, 'SERVICES', self.services)
 
-    def get_actual(self) -> Mapping[str, Event]:
+    async def run_services(self):
         """
-        Get actual channel values.
+        Run all services.
         """
-        return {key: Event(**value) for key, value in self.db[ACTUAL_KEY].items()}
+        logger.info('Running services…')
+        try:
+            await gather(*[self.run_service(service) for service in self.services])
+        except CancelledError:
+            for service in self.services:
+                with suppress(Exception):
+                    logger.info('Closing {}…', service)
+                    await service.close()
+        except Exception as e:
+            logger.opt(exception=e).critical('Failed to run services.')
 
-    def get_log(self, channel: str, period: timedelta) -> List[Event]:
+    async def run_service(self, service: Service):
         """
-        Gets the channel log within the specified period until now.
+        Run the single service.
         """
-        return [
-            Event(**event)
-            for event in self.db[f'log:{channel}'][timestamp_key(datetime.now() - period):]
-        ]
+        n_errors = 0
+        while True:
+            logger.info('Running {}…', service)
+            try:
+                async for event in service.events:
+                    n_errors = 0  # the service successfully generated an event
+                    await self.on_event(event)
+            except CancelledError:
+                logger.info('Stopped service {}.', service)
+                break
+            except Exception as e:
+                n_errors += 1
+                if isinstance(e, (ConnectionError, ClientConnectorError, asyncio.TimeoutError)):
+                    logger.error('{} has raised a connection error: {}', service, e)
+                else:
+                    logger.opt(exception=e).error('{} has failed.', service)
+            else:
+                n_errors = 0  # the service has finished normally
+            if n_errors:
+                delay = 2.0 ** min(n_errors, 16)
+                logger.error('{} errors. Restarting in {}…', n_errors, timedelta(seconds=delay))
+                await sleep(delay)
 
-    async def trigger_event(self, event: Event):
+    async def on_event(self, event: Event):
         """
-        Handle a single event in the application context.
+        Handle the single event.
         """
         logger.info('{key} = {value!r}', key=event.channel, value=event.value)
         previous = self.db[ACTUAL_KEY].get(event.channel)
-
-        await setattr_async(self.db[ACTUAL_KEY], event.channel, event.dict(exclude=EVENT_EXCLUDE))
-        if event.is_logged:
-            collection = self.db[f'log:{event.channel}']
-            await setattr_async(collection, timestamp_key(event.timestamp), event.dict(include=EVENT_INCLUDE))
-
+        await self.save_event(event)
         previous = Event(**previous) if previous is not None else None
         # noinspection PyAsyncCall
         create_task(self.router.on_event(
@@ -76,6 +101,30 @@ class Context:
             actual=self.get_actual(),
             session=self.session,
         ))
+
+    @run_in_executor
+    def save_event(self, event: Event):
+        with self.db:
+            self.db[ACTUAL_KEY][event.channel] = event.dict(exclude=EVENT_EXCLUDE)
+            if event.is_logged:
+                self.db[f'log:{event.channel}'][timestamp_key(event.timestamp)] = event.dict(include=EVENT_INCLUDE)
+
+    # TODO: perhaps move to `db_utils`.
+    def get_actual(self) -> Mapping[str, Event]:
+        """
+        Get actual channel values.
+        """
+        return {key: Event(**value) for key, value in self.db[ACTUAL_KEY].items()}
+
+    # TODO: perhaps move to `db_utils`.
+    def get_log(self, channel: str, period: timedelta) -> List[Event]:
+        """
+        Gets the channel log within the specified period until now.
+        """
+        return [
+            Event(**event)
+            for event in self.db[f'log:{channel}'][timestamp_key(datetime.now() - period):]
+        ]
 
     async def close(self):
         self.db.close()