+++ my_iot/helpers.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from asyncio import get_running_loop
 from datetime import datetime, timezone
-from typing import Any, Callable, TypeVar
+from functools import partial, wraps
+from typing import Awaitable, Callable, TypeVar
 
 T = TypeVar('T')
 utc = timezone.utc
@@ -16,9 +17,8 @@ def timestamp_key(timestamp: datetime) -> str:
     return timestamp.astimezone(utc).strftime('%Y%m%d%H%M%S%f')
 
 
-async def run_in_executor(callable_: Callable[..., T], *args: Any) -> T:
-    return get_running_loop().run_in_executor(None, callable_, *args)
-
-
-async def setattr_async(object_: Any, attribute: Any, value: Any):
-    await run_in_executor(setattr, object_, attribute, value)
+def run_in_executor(callable_: Callable[..., T]) -> Callable[..., Awaitable[T]]:
+    @wraps(callable_)
+    async def wrapper(*args, **kwargs) -> T:
+        return await get_running_loop().run_in_executor(None, partial(callable_, *args, **kwargs))
+    return wrapper